fix: PRレビュー対応（fetchTastesエラー修正・静的データ削除・フィルター関東絞り込み）
