+++ frontend/src/app/lp/_components/PricingSection/pricingWithToggle.tsx
@@ -1,14 +1,15 @@
 "use client";
 
 import { useState } from "react";
-import { getPlansForWeight, type WeightGrams } from "../../_lib/planData";
+import type { PlanDetail, WeightGrams } from "../../_lib/planApi";
 import WeightToggle from "../WeightToggle/weightToggle";
 import PricingCarousel from "./pricingCarousel";
 import styles from "./pricingSection.module.css";
 
 export default function PricingWithToggle() {
   const [weight, setWeight] = useState<WeightGrams>(30);
-  const filteredPlans = getPlansForWeight(weight);
+  // NOTE: このコンポーネントは現在未使用。プランデータはAPIから取得するため空配列を渡す。
+  const filteredPlans: PlanDetail[] = [];
 
   return (
     <>
